requirements.txt
aiohappyeyeballs                 2.4.3
aiohttp                          3.9.5
aiosignal                        1.3.1
annotated-types                  0.7.0
anyio                            3.7.1
apimatic-core                    0.2.16
apimatic-core-interfaces         0.1.5
apimatic-requests-client-adapter 0.1.6
async-timeout                    4.0.3
attrs                            24.2.0
blinker                          1.9.0
CacheControl                     0.12.14
certifi                          2024.8.30
cffi                             1.17.1
charset-normalizer               3.4.0
click                            8.1.7
colorama                         0.4.6
cryptography                     44.0.0
dataclasses-json                 0.6.7
deprecation                      2.1.0
distro                           1.9.0
exceptiongroup                   1.2.2
Flask                            3.1.0
Flask-Cors                       5.0.0
frozenlist                       1.5.0
greenlet                         3.1.1
h11                              0.14.0
httpcore                         1.0.7
httpx                            0.27.2
httpx-sse                        0.4.0
idna                             3.10
itsdangerous                     2.2.0
Jinja2                           3.1.4
jiter                            0.8.0
jsonpatch                        1.33
jsonpickle                       3.0.4
jsonpointer                      2.4
langchain                        0.3.9
langchain-community              0.3.8
langchain-core                   0.3.21
langchain-openai                 0.2.0
langchain-pinecone               0.2.0
langchain-text-splitters         0.3.2
langsmith                        0.1.147
MarkupSafe                       3.0.2
marshmallow                      3.23.1
msgpack                          1.1.0
multidict                        6.1.0
mypy-extensions                  1.0.0
numpy                            1.26.4
openai                           1.7.2
orjson                           3.10.12
packaging                        24.2
pdfminer.six                     20231228
pdfplumber                       0.11.4
pillow                           11.0.0
pinecone                         5.4.1
pinecone-client                  5.0.1
pinecone-plugin-inference        1.1.0
pinecone-plugin-interface        0.0.7
pip                              21.2.3
propcache                        0.2.0
pycparser                        2.22
pydantic                         2.10.2
pydantic_core                    2.27.1
pydantic-settings                2.6.1
pypdf                            5.1.0
pypdfium2                        4.30.0
python-dateutil                  2.8.2
python-dotenv                    1.0.1
PyYAML                           6.0.2
regex                            2024.11.6
requests                         2.32.3
requests-toolbelt                1.0.0
setuptools                       75.6.0
six                              1.16.0
sniffio                          1.3.1
SQLAlchemy                       2.0.35
squareup                         39.0.0.20241120
tenacity                         9.0.0
tiktoken                         0.8.0
tqdm                             4.67.1
typing_extensions                4.12.2
typing-inspect                   0.9.0
urllib3                          2.2.3
Werkzeug                         3.1.3
yarl                             1.18.0